New experiment batch
