+++ examples/AAAI_expetiments/results/SOD_NGD_30.csv
@@ -1,0 +1,4 @@
+,grid_res,error_train_pressure,error_train_velocity,error_train_density,error_test_pressure,error_test_velocity,error_test_density,loss,time,type
+0,10,0.18930421769618988,0.3853941559791565,0.15942026674747467,0.18930421769618988,0.3853941559791565,0.15942026674747467,1.70699004229391e-05,7087.670613765717,burgers1d_NGD
+1,20,0.15751603245735168,0.37559136748313904,0.13914063572883606,0.15751603245735168,0.37559136748313904,0.13914063572883606,6.364621367538348e-05,34442.87049126625,burgers1d_NGD
+2,30,0.10488904267549515,0.24950696527957916,0.09405776858329773,0.10488904267549515,0.24950696527957916,0.09405776858329773,0.0003709829179570079,126339.88511753082,burgers1d_NGD
