+++ examples/AAAI_expetiments/results/SOD_lam_60.csv
@@ -1,0 +1,7 @@
+,grid_res,error_train_pressure,error_train_velocity,error_train_density,error_test_pressure,error_test_velocity,error_test_density,loss,time,type
+0,10,0.1566501259803772,0.3157793879508972,0.18498679995536804,0.1566501259803772,0.3157793879508972,0.18498679995536804,0.02738380804657936,1566.9964413642883,SOD_lam
+1,20,0.08857022225856781,0.20835207402706146,0.1412087082862854,0.08857022225856781,0.20835207402706146,0.1412087082862854,53.78704833984375,4726.667871952057,SOD_lam
+2,30,0.07393793016672134,0.2053108960390091,0.10261039435863495,0.07393793016672134,0.2053108960390091,0.10261039435863495,11.258476257324219,19569.039818525314,SOD_lam
+3,40,0.06260708719491959,0.17495088279247284,0.07783371210098267,0.06260708719491959,0.17495088279247284,0.07783371210098267,0.030993333086371422,27665.415113210678,SOD_lam
+4,50,0.05271263048052788,0.17061816155910492,0.07062949985265732,0.05271263048052788,0.17061816155910492,0.07062949985265732,0.03649904951453209,67399.78881192207,SOD_lam
+5,60,0.049207136034965515,0.17614200711250305,0.06759429723024368,0.049207136034965515,0.17614200711250305,0.06759429723024368,0.03441302105784416,128703.06282138824,SOD_lam
